Move gamry from parsers repositiory to shared parsers/schema repository
