+++ tests/data/test.example-format.txt
@@ -1,0 +1,4 @@
+# Hello World!
+1 0 1
+1 1 1
+1 0 1
